fix merge issues

+++ client/src/main.rs
@@ -531,9 +531,13 @@ fn run_app() -> Result<(), Box<dyn std::error::Error>> {
                 if let Some((w, h, shared_handle, format, uv_y_offset, _ts)) = native_frame {
                     frame_consumed.store(true, std::sync::atomic::Ordering::Release);
                     if let Some(presenter) = native_stream_presenter.as_mut() {
-                        if let Err(err) =
-                            presenter.render_shared_frame(w, h, shared_handle, format, uv_y_offset)
-                        {
+                        if let Err(err) = presenter.render_shared_frame(
+                            w,
+                            h,
+                            shared_handle,
+                            format as u32,
+                            uv_y_offset,
+                        ) {
                             log::warn!("Native shared-frame render failed: {}", err);
                         } else {
                             frame_timer_consumed = frame_timer_consumed.saturating_add(1);
